+++ src/app/formulario/participante/page.tsx
@@ -1,792 +1,453 @@
 'use client';
-import { useState, useMemo, useEffect } from 'react';
+import { useState, useEffect } from 'react';
+import { useRouter } from 'next/navigation';
 import InputField from '@components/buttons_inputs/InputField';
-import FiltroEvento from '@/components/headers_menu_users/FiltroEvento';
 import Dropdown from '@components/buttons_inputs/Dropdown';
-import PageTitle from '@/components/headers_menu_users/pageTitle';
 import Button from '@components/buttons_inputs/Button';
-import Pagination from '@/components/buttons_inputs/Pagination';
 import Checkbox from '@components/buttons_inputs/Checkbox';
 import withIconDecorator from '@/components/decorators/IconDecorator';
-import { FlowerLotus, User, AddressBook, SketchLogo, Check, Eye, Star, Megaphone, X, UserSound, ChatTeardropText, Grains, Student, GraduationCap, BookOpenText, SealWarning, Heart, FilePdf, BookmarksSimple } from '@/components/icons';
-
-// Import icons using the specified path (placeholders)
-import Location from '@components/icons/Gps'; // For localización field
-import Send from '@components/icons/ArrowFatRight'; // For submit button
+import { Modal, Toast } from '@/components/buttons_inputs/FormNotification';
+import { FlowerLotus, AddressBook, X, User, Phone, Envelope, GraduationCap, Flag } from '@components/icons';
+import Send from '@components/icons/ArrowFatRight';
 import Navbar from '@/components/headers_menu_users/navbar';
+import ParticipantGroupSelectionTable from '@/components/tables/GroupSelectionTable';
 
-interface Coordinator {
-    name: string;
-    lastNameP: string;
-    lastNameM: string;
-    email: string;
-    phone: string;
+interface Group {
+  id_group: number;
+  name: string;
+  mode: string;
+  sede: string;
+  cupo: string;
+  horarios: string;
+  fechas: string;
 }
 
-interface GeneralCoordinator extends Coordinator {
-    gender: string;
-    username: string;
-    password: string;
-    confirmPassword: string;
+interface ParticipantData {
+  name: string;
+  paternal_name: string;
+  maternal_name: string;
+  email: string;
+  year: string;
+  education: string;
 }
 
-interface Venue {
-    name: string;
-    location: string;
-    address: string;
-}
-
-interface Grupos {
-    grupo: string;
-    modalidad: string;
-    sede: string;
-    cupo: string;
-    horarios: string;
-    fechas: string;
+interface TutorData {
+  name: string;
+  paternal_name: string;
+  maternal_name: string;
+  email: string;
+  phone_number: string;
 }
 
 interface FormData {
-    generalCoordinator: GeneralCoordinator;
-    associatedCoordinator: Coordinator;
-    staffCoordinator: Coordinator;
-    participantsCoordinator: Coordinator;
-    venue: Venue;
+  participant: ParticipantData;
+  tutor: TutorData;
+  participation_file: File | null;
+  preferred_group: number | null;
+  privacy_accepted: boolean;
 }
 
+// Define sections that are objects for handleInputChange
+type ObjectSections = 'participant' | 'tutor';
+type ObjectSectionData = ParticipantData | TutorData;
+type ObjectSectionKeys = keyof ParticipantData | keyof TutorData;
+
 const ParticipantRegistrationForm: React.FC = () => {
-    const [formData, setFormData] = useState<FormData>({
-        generalCoordinator: {
-            name: '',
-            lastNameP: '',
-            lastNameM: '',
-            email: '',
-            phone: '',
-            gender: 'Mujer',
-            username: '',
-            password: '',
-            confirmPassword: '',
+  const router = useRouter();
+  const [formData, setFormData] = useState<FormData>({
+    participant: {
+      name: '',
+      paternal_name: '',
+      maternal_name: '',
+      email: '',
+      year: '',
+      education: '',
+    },
+    tutor: {
+      name: '',
+      paternal_name: '',
+      maternal_name: '',
+      email: '',
+      phone_number: '',
+    },
+    participation_file: null,
+    preferred_group: null,
+    privacy_accepted: false,
+  });
+
+  const [groups, setGroups] = useState<Group[]>([]);
+  const [errors, setErrors] = useState<string[]>([]);
+  const [success, setSuccess] = useState<string | null>(null);
+  const [isErrorModalOpen, setIsErrorModalOpen] = useState(false);
+  const [isSuccessToastOpen, setIsSuccessToastOpen] = useState(false);
+
+  // Fetch groups from API
+  useEffect(() => {
+    const fetchGroups = async () => {
+      try {
+        const response = await fetch('http://localhost:3000/api/groups');
+        const data = await response.json();
+        const transformedGroups = data.map((group: any) => ({
+          id_group: group.id_group,
+          name: group.name,
+          mode: group.mode || 'Presencial',
+          sede: group.venues?.name || 'N/A',
+          cupo: `${group.occupied_places || 0}/${group.max_places || 'N/A'} Personas`,
+          horarios: `${group.start_hour || 'N/A'} - ${group.end_hour || 'N/A'}`,
+          fechas: `${group.start_date ? new Date(group.start_date).toLocaleDateString('es-MX') : 'N/A'} - ${group.end_date ? new Date(group.end_date).toLocaleDateString('es-MX') : 'N/A'}`,
+        }));
+        setGroups(transformedGroups);
+      } catch (err) {
+        setErrors(['Error al cargar los grupos']);
+        setIsErrorModalOpen(true);
+      }
+    };
+    fetchGroups();
+  }, []);
+
+  const handleInputChange = <T extends ObjectSections>(
+    section: T,
+    field: ObjectSectionKeys,
+    value: string
+  ) => {
+    setFormData((prev) => ({
+      ...prev,
+      [section]: {
+        ...prev[section],
+        [field]: value,
+      },
+    }));
+  };
+
+  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const file = e.target.files?.[0];
+    if (file && file.type === 'application/pdf') {
+      setFormData((prev) => ({ ...prev, participation_file: file }));
+    } else {
+      setErrors(['El archivo debe ser un PDF']);
+      setIsErrorModalOpen(true);
+    }
+  };
+
+  const handleGroupSelect = (id_group: number) => {
+    setFormData((prev) => ({ ...prev, preferred_group: id_group }));
+  };
+
+  const validateForm = () => {
+    const newErrors: string[] = [];
+    if (!formData.participant.name) newErrors.push('El nombre es obligatorio');
+    if (!formData.participant.paternal_name) newErrors.push('El apellido paterno es obligatorio');
+    if (!formData.participant.email || !/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(formData.participant.email))
+      newErrors.push('El correo electrónico debe ser válido');
+    if (!formData.participant.year) newErrors.push('El grado es obligatorio');
+    if (!formData.participant.education) newErrors.push('La escolaridad es obligatoria');
+    if (!formData.participation_file) newErrors.push('El archivo de participación es obligatorio');
+    if (!formData.tutor.name) newErrors.push('El nombre del tutor es obligatorio');
+    if (!formData.tutor.paternal_name) newErrors.push('El apellido paterno del tutor es obligatorio');
+    if (!formData.tutor.email || !/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(formData.tutor.email))
+      newErrors.push('El correo del tutor debe ser válido');
+    if (!formData.tutor.phone_number) newErrors.push('El celular del tutor es obligatorio');
+    if (!formData.privacy_accepted) newErrors.push('Debes aceptar el aviso de privacidad');
+    return newErrors;
+  };
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+
+    const validationErrors = validateForm();
+    if (validationErrors.length > 0) {
+      setErrors(validationErrors);
+      setIsErrorModalOpen(true);
+      setSuccess(null);
+      setIsSuccessToastOpen(false);
+      return;
+    }
+
+    const formDataToSend = new FormData();
+    formDataToSend.append('name', formData.participant.name);
+    formDataToSend.append('paternal_name', formData.participant.paternal_name);
+    formDataToSend.append('maternal_name', formData.participant.maternal_name);
+    formDataToSend.append('email', formData.participant.email);
+    formDataToSend.append('year', formData.participant.year);
+    formDataToSend.append('education', formData.participant.education);
+    formDataToSend.append('tutor[name]', formData.tutor.name);
+    formDataToSend.append('tutor[paternal_name]', formData.tutor.paternal_name);
+    formDataToSend.append('tutor[maternal_name]', formData.tutor.maternal_name);
+    formDataToSend.append('tutor[email]', formData.tutor.email);
+    formDataToSend.append('tutor[phone_number]', formData.tutor.phone_number);
+    if (formData.preferred_group) {
+      formDataToSend.append('preferred_group', formData.preferred_group.toString());
+    }
+    if (formData.participation_file) {
+      formDataToSend.append('participation_file', formData.participation_file);
+    }
+
+    const token = localStorage.getItem('token');
+    try {
+      const response = await fetch('http://localhost:3000/api/participants', {
+        method: 'POST',
+        headers: {
+          Authorization: `Bearer ${token}`,
         },
-        associatedCoordinator: {
-            name: '',
-            lastNameP: '',
-            lastNameM: '',
-            email: '',
-            phone: '',
-        },
-        staffCoordinator: {
-            name: '',
-            lastNameP: '',
-            lastNameM: '',
-            email: '',
-            phone: '',
-        },
-        participantsCoordinator: {
-            name: '',
-            lastNameP: '',
-            lastNameM: '',
-            email: '',
-            phone: '',
-        },
-        venue: {
-            name: '',
-            location: 'Puebla',
-            address: '',
-        },
-    });
+        body: formDataToSend,
+      });
 
-    const [profileImage, setProfileImage] = useState<File | null>(null);
-    const [logo, setLogo] = useState<File | null>(null);
-    const [participationFile, setParticipationFile] = useState<File | null>(null);
-    const [privacyAccepted, setPrivacyAccepted] = useState(false);
-    const [showPassword, setShowPassword] = useState(false);
-    const [showConfirmPassword, setShowConfirmPassword] = useState(false);
-    const [errors, setErrors] = useState<string[]>([]);
-    const [success, setSuccess] = useState<string | null>(null);
-    const [inputValue, setInputValue] = useState('');
-    const [section, setSection] = useState('__All__'); // Inicializar con "Todas"
-    const [currentPage, setCurrentPage] = useState(0);
-    const [isPopupOpen, setIsPopupOpen] = useState(false); // Estado para controlar el popup
-    const [selectedSede, setSelectedSede] = useState<Grupos | null>(null); // Sede seleccionada para eliminar
-    const [chosenSede, setChosenSede] = useState<string | null>(null); // Nueva state para la sede elegida
+      const data = await response.json();
 
-    const [chosenGrupo, setChosenGrupo] = useState<string | null>(null); // Nueva state para la sede elegida
-
-    const DecoratedFlowerLotus = withIconDecorator(FlowerLotus);
-    const DecoratedSealWarning = withIconDecorator(SealWarning);
-    const DecoratedAddressBook = withIconDecorator(AddressBook);
-    const DecoratedStudent = withIconDecorator(Student);
-    const DecoratedBookmarksSimple = withIconDecorator(BookmarksSimple);
-    const DecoratedHeart = withIconDecorator(Heart);
-    const DecoratedFilePdf = withIconDecorator(FilePdf);
-    const DecoratedMegaphone = withIconDecorator(Megaphone);
-
-    // Handle input changes
-    type GeneralCoordinatorKeys = keyof GeneralCoordinator; // 'name' | 'lastNameP' | 'lastNameM' | 'email' | 'phone' | 'gender' | 'username' | 'password' | 'confirmPassword'
-    type CoordinatorKeys = keyof Coordinator; // 'name' | 'lastNameP' | 'lastNameM' | 'email' | 'phone'
-    type VenueKeys = keyof Venue; // 'name' | 'location' | 'address'
-
-    type Section = keyof FormData; // 'generalCoordinator' | 'associatedCoordinator' | 'staffCoordinator' | 'participantsCoordinator' | 'venue'
-
-    type SubSectionMap = {
-        generalCoordinator: GeneralCoordinatorKeys;
-        associatedCoordinator: CoordinatorKeys;
-        staffCoordinator: CoordinatorKeys;
-        participantsCoordinator: CoordinatorKeys;
-        venue: VenueKeys;
-    };
-
-    const rowsPerPage = 10;
-
-    const gruposData: Grupos[] = [
-        { sede: 'ITESM Monterrey', modalidad: 'Presencial', grupo: 'Sol', cupo: '10 Personas', fechas: 'DD/MM - DD/MM', horarios: 'HH:MM - HH:MM' },
-        { sede: 'ITESM Puebla', modalidad: 'Presencial', grupo: 'Luna', cupo: '12 Personas', fechas: 'DD/MM - DD/MM', horarios: 'HH:MM - HH:MM' },
-        { sede: 'ITESM Guadalajara', modalidad: 'En línea', grupo: 'Montaña', cupo: '8 Personas', fechas: 'DD/MM - DD/MM', horarios: 'HH:MM - HH:MM' },
-        { sede: 'ITESM Toluca', modalidad: 'Presencial', grupo: 'Mar', cupo: 'Ximena 11', fechas: 'DD/MM - DD/MM', horarios: 'HH:MM - HH:MM' },
-        { sede: 'ITESM Querétaro', modalidad: 'En línea', grupo: 'Sol', cupo: '2 Personas', fechas: 'DD/MM - DD/MM', horarios: 'HH:MM - HH:MM' },
-        { sede: 'ITESM Cuernavaca', modalidad: 'Presencial', grupo: 'Luna', cupo: '10 Personas', fechas: 'DD/MM - DD/MM', horarios: 'HH:MM - HH:MM' },
-        { sede: 'ITESM Hidalgo', modalidad: 'Presencial', grupo: 'Montaña', cupo: '7 Personas', fechas: 'DD/MM - DD/MM', horarios: 'HH:MM - HH:MM' },
-    ];
-
-    // Obtener modalidades únicas
-    const uniqueModalidades = Array.from(new Set(gruposData.map(sedes => sedes.modalidad))).sort();
-    const modalidadesOptions = [
-        { label: 'Todas', value: '__All__' },
-        ...uniqueModalidades.map(modalidad => ({ label: modalidad, value: modalidad })),
-    ];
-
-    // Filtrar los datos según el valor de búsqueda y sede
-    const filteredData = useMemo(() => {
-        const searchTerm = inputValue.toLowerCase().trim();
-        return gruposData.filter(sedes => {
-            // Filtro por nombre de Sede
-            const matchesSearch = !searchTerm || sedes.sede.toLowerCase().includes(searchTerm);
-
-            // Filtro por sede
-            const matchesSede = section === '__All__' ? true : sedes.modalidad === section;
-
-            return matchesSearch && matchesSede;
-        });
-    }, [inputValue, section]);
-
-    const totalPages = Math.ceil(filteredData.length / rowsPerPage);
-    const paginatedData = filteredData.slice(currentPage * rowsPerPage, (currentPage + 1) * rowsPerPage);
-
-    // Añadimos un useEffect para reiniciar currentPage cuando filteredData cambie
-    useEffect(() => {
-        // Si la página actual es mayor o igual al número total de páginas después del filtrado,
-        // ajustamos currentPage para que no exceda el rango válido
-        if (currentPage >= totalPages && totalPages > 0) {
-            setCurrentPage(totalPages - 1);
-        } else if (totalPages === 0) {
-            setCurrentPage(0);
+      if (!response.ok) {
+        if (response.status === 422 && data.errors) {
+          const backendErrors = data.errors.map((err: any) => err.msg);
+          throw new Error(backendErrors.join(', '));
         }
-    }, [filteredData.length, currentPage, totalPages]);
+        throw new Error(data.message || 'Error al registrar participante');
+      }
 
-    const sectionFilterChange = (value: string) => {
-        setSection(value);
-        setInputValue(''); // Resetear búsqueda al cambiar de sección
-        setCurrentPage(0); // Resetear página al cambiar de filtro
-    };
+      setSuccess(data.message || 'Participante registrado exitosamente');
+      setIsSuccessToastOpen(true);
+      setErrors([]);
+      setIsErrorModalOpen(false);
+      setFormData({
+        participant: { name: '', paternal_name: '', maternal_name: '', email: '', year: '', education: '' },
+        tutor: { name: '', paternal_name: '', maternal_name: '', email: '', phone_number: '' },
+        participation_file: null,
+        preferred_group: null,
+        privacy_accepted: false,
+      });
+    } catch (err: any) {
+      setErrors([err.message]);
+      setIsErrorModalOpen(true);
+      setSuccess(null);
+      setIsSuccessToastOpen(false);
+    }
+  };
 
-    // Función para abrir el popup de confirmación y seleccionar sede
-    const handleSelectClick = (sedes: Grupos) => {
-        setSelectedSede(sedes);
-        setChosenSede(sedes.sede); // Actualizar la sede elegida al hacer click en "Check"
-        setChosenGrupo(sedes.grupo); // Actualizar la sede elegida al hacer click en "Check"
-        setIsPopupOpen(true);
-    };
+  return (
+    <form className="min-h-screen bg-gray-900 text-white p-4 md:p-8 flex flex-col items-center" onSubmit={handleSubmit}>
+      <Navbar />
+      <div className="w-full max-w-6xl bg-gray-800 rounded-lg shadow-lg p-6 md:p-8">
+        {/* Header */}
+        <div className="flex justify-between items-center mb-6">
+          <div className="flex items-center">
+            <div className="w-2 h-16 mr-4 bg-purple-600 rounded-full"></div>
+            <h1 className="text-2xl">
+              <span className="italic">Formulario de Registro</span><br />
+              <span className="font-bold text-3xl">Participantes</span>
+            </h1>
+          </div>
+          <Button
+            label=""
+            variant="error"
+            round
+            showLeftIcon
+            IconLeft={X}
+            onClick={() => router.push('/')}
+            className="px-4 py-2 rounded-full flex items-center"
+          />
+        </div>
 
-    // Función para cerrar el popup
-    const handleClosePopup = () => {
-        setIsPopupOpen(false);
-        setSelectedSede(null);
-    };
+        {/* Datos del Participante */}
+        <div className="mb-6">
+          <h2 className="text-xl md:text-2xl font-semibold flex items-center mb-2">
+            <span className="mr-2"><FlowerLotus /></span> Datos del Participante
+          </h2>
+          <p className="text-gray-400 text-sm md:text-base mb-4">
+            Responde con sinceridad las siguientes preguntas. Campos con (*) son obligatorios.
+          </p>
+        </div>
 
-    const handleInputChange = <S extends Section>(
-        section: S,
-        field: string,
-        value: string,
-        subSection?: SubSectionMap[S]
-    ) => {
-        setFormData((prev: FormData) => {
-            if (subSection) {
-                return {
-                    ...prev,
-                    [section]: {
-                        ...prev[section],
-                        [subSection]: {
-                            ...prev[section][subSection as keyof typeof prev[S]], // Type assertion for subSection
-                            [field]: value,
-                        },
-                    },
-                };
-            }
-            return {
-                ...prev,
-                [section]: {
-                    ...prev[section],
-                    [field]: value,
-                },
-            };
-        });
-    };
+        <div className="grid grid-cols-1 md:grid-cols-3 gap-4 mb-6">
+          <InputField
+            label="Nombre(s)*"
+            placeholder="Edna"
+            variant="primary"
+            icon="User"
+            value={formData.participant.name}
+            onChangeText={(value) => handleInputChange('participant', 'name', value)}
+          />
+          <InputField
+            label="Apellido Paterno*"
+            placeholder="Moda"
+            variant="primary"
+            icon="User"
+            value={formData.participant.paternal_name}
+            onChangeText={(value) => handleInputChange('participant', 'paternal_name', value)}
+          />
+          <InputField
+            label="Apellido Materno"
+            placeholder="Apellido"
+            variant="primary"
+            icon="User"
+            value={formData.participant.maternal_name}
+            onChangeText={(value) => handleInputChange('participant', 'maternal_name', value)}
+          />
+          <InputField
+            label="Correo Electrónico*"
+            placeholder="ednamoda@disney.com"
+            variant="accent"
+            icon="Envelope"
+            value={formData.participant.email}
+            onChangeText={(value) => handleInputChange('participant', 'email', value)}
+          />
+          <Dropdown
+            label="Grado*"
+            options={['1º', '2º', '3º']}
+            value={formData.participant.year}
+            onChange={(value) => handleInputChange('participant', 'year', value)}
+            variant="accent"
+            Icon={withIconDecorator(GraduationCap)}
+          />
+          <Dropdown
+            label="Escolaridad*"
+            options={['Secundaria', 'Preparatoria']}
+            value={formData.participant.education}
+            onChange={(value) => handleInputChange('participant', 'education', value)}
+            variant="accent"
+            Icon={withIconDecorator(GraduationCap)}
+          />
+        </div>
 
-    // Handle file uploads
-    const handleFileChange = (
-        e: React.ChangeEvent<HTMLInputElement>,
-        setFile: React.Dispatch<React.SetStateAction<File | null>>
-    ) => {
-        const file = e.target.files?.[0];
-        if (file) {
-            setFile(file);
-        }
-    };
+        {/* Datos del Tutor */}
+        <div className="mb-6">
+          <h2 className="text-xl md:text-2xl font-semibold flex items-center mb-2">
+            <span className="mr-2"><FlowerLotus /></span> Datos del Tutor
+          </h2>
+          <p className="text-gray-400 text-sm md:text-base mb-4">
+            Pídele a tu tutor que llene esta sección. Campos con (*) son obligatorios.
+          </p>
+        </div>
 
-    // Convert file to base64
-    const fileToBase64 = (file: File): Promise<string> => {
-        return new Promise((resolve, reject) => {
-            const reader = new FileReader();
-            reader.readAsDataURL(file);
-            reader.onload = () => {
-                const base64String = (reader.result as string).split(',')[1]; // Remove the "data:..." prefix
-                resolve(base64String);
-            };
-            reader.onerror = (error) => reject(error);
-        });
-    };
+        <div className="grid grid-cols-1 md:grid-cols-3 gap-4 mb-6">
+          <InputField
+            label="Nombre(s)*"
+            placeholder="Roberto"
+            variant="primary"
+            icon="User"
+            value={formData.tutor.name}
+            onChangeText={(value) => handleInputChange('tutor', 'name', value)}
+          />
+          <InputField
+            label="Apellido Paterno*"
+            placeholder="López"
+            variant="primary"
+            icon="User"
+            value={formData.tutor.paternal_name}
+            onChangeText={(value) => handleInputChange('tutor', 'paternal_name', value)}
+          />
+          <InputField
+            label="Apellido Materno"
+            placeholder="Juárez"
+            variant="primary"
+            icon="User"
+            value={formData.tutor.maternal_name}
+            onChangeText={(value) => handleInputChange('tutor', 'maternal_name', value)}
+          />
+          <InputField
+            label="Correo Electrónico*"
+            placeholder="tutor@example.com"
+            variant="accent"
+            icon="Envelope"
+            value={formData.tutor.email}
+            onChangeText={(value) => handleInputChange('tutor', 'email', value)}
+          />
+          <InputField
+            label="Celular*"
+            placeholder="+522221234567"
+            variant="accent"
+            icon="Phone"
+            value={formData.tutor.phone_number}
+            onChangeText={(value) => handleInputChange('tutor', 'phone_number', value)}
+          />
+        </div>
 
-    // Client-side validation
-    const validateForm = () => {
-        const newErrors: string[] = [];
+        {/* Selección de Grupo */}
+        <div className="mb-6">
+          <h2 className="text-xl md:text-2xl font-semibold flex items-center mb-2">
+            <span className="mr-2"><AddressBook /></span> Selección de Grupo
+          </h2>
+          <p className="text-gray-400 text-sm md:text-base mb-4">
+            Selecciona el grupo que prefieres. Usa los botones para más detalles.
+          </p>
+        </div>
 
-        // Venue validation
-        if (!formData.venue.name) newErrors.push('El nombre de la SEDE es obligatorio');
-        if (!formData.venue.location) newErrors.push('La localización de la SEDE es obligatoria');
-        if (!formData.venue.address) newErrors.push('La dirección de la SEDE es obligatoria');
-        if (!participationFile) newErrors.push('El archivo de participación es obligatorio');
+        <div className="flex justify-center items-center mx-auto w-[80%] h-[50px] rounded-t-[15px] mt-8 bg-gray-700 text-xs sm:text-base md:text-lg lg:text-xl">
+          <h2 className="mx-4 font-semibold">Grupo Elegido: </h2>
+          <p>{groups.find(g => g.id_group === formData.preferred_group)?.name || 'Ninguno'}</p>
+        </div>
+        <ParticipantGroupSelectionTable
+          onSelect={handleGroupSelect}
+          selectedGroupId={formData.preferred_group ?? undefined}
+          rowsPerPage={4}
+        />
 
-        // General Coordinator validation
-        if (!formData.generalCoordinator.name)
-            newErrors.push('El nombre de la coordinadora general es obligatorio');
-        if (!formData.generalCoordinator.lastNameP)
-            newErrors.push('El apellido paterno de la coordinadora general es obligatorio');
-        if (!formData.generalCoordinator.email)
-            newErrors.push('El correo electrónico de la coordinadora general es obligatorio');
-        else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(formData.generalCoordinator.email))
-            newErrors.push('El correo electrónico de la coordinadora general debe ser válido');
-        if (!formData.generalCoordinator.phone)
-            newErrors.push('El celular de la coordinadora general es obligatorio');
-        if (!formData.generalCoordinator.gender)
-            newErrors.push('El sexo de la coordinadora general es obligatorio');
-        if (!formData.generalCoordinator.username)
-            newErrors.push('El nombre de usuario de la coordinadora general es obligatorio');
-        if (!formData.generalCoordinator.password)
-            newErrors.push('La contraseña de la coordinadora general es obligatoria');
-        else {
-            if (formData.generalCoordinator.password.length < 8)
-                newErrors.push('La contraseña debe tener al menos 8 caracteres');
-            if (!/[A-Z]/.test(formData.generalCoordinator.password))
-                newErrors.push('La contraseña debe contener al menos una mayúscula');
-            if (!/[a-z]/.test(formData.generalCoordinator.password))
-                newErrors.push('La contraseña debe contener al menos una minúscula');
-            if (!/[!@#$%^&*(),.?":{}|<>]/.test(formData.generalCoordinator.password))
-                newErrors.push('La contraseña debe contener al menos un carácter especial');
-        }
-        if (formData.generalCoordinator.password !== formData.generalCoordinator.confirmPassword)
-            newErrors.push('Las contraseñas no coinciden');
+        {/* Permiso de Participación */}
+        <div className="mt-8">
+          <h2 className="text-xl md:text-2xl font-semibold flex items-center mb-2">
+            <span className="mr-2"><FlowerLotus /></span> Permiso de Participación
+          </h2>
+          <p className="text-gray-400 text-sm md:text-base mb-4">
+            Sube tu permiso firmado (PDF obligatorio).
+          </p>
+          <div className="p-4 bg-white text-black rounded-lg">
+            <input
+              type="file"
+              accept="application/pdf"
+              onChange={handleFileChange}
+              className="mt-4"
+            />
+            {formData.participation_file && (
+              <p className="mt-2 text-xs text-gray-600">
+                Archivo seleccionado: {formData.participation_file.name}
+              </p>
+            )}
+          </div>
+        </div>
 
-        // Associated Coordinator validation
-        if (formData.associatedCoordinator.name) {
-            if (!formData.associatedCoordinator.email)
-                newErrors.push('El correo electrónico de la coordinadora asociada es obligatorio');
-            else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(formData.associatedCoordinator.email))
-                newErrors.push('El correo electrónico de la coordinadora asociada debe ser válido');
-            if (!formData.associatedCoordinator.phone)
-                newErrors.push('El celular de la coordinadora asociada es obligatorio');
-        }
+        {/* Aviso de Privacidad */}
+        <div className="mt-8">
+          <h2 className="text-xl md:text-2xl font-semibold flex items-center mb-2">
+            <span className="mr-2"><FlowerLotus /></span> Aviso de Privacidad
+          </h2>
+          <p className="text-gray-400 text-sm">
+            Confirma que has leído y aceptas el aviso de privacidad:
+            <a href="https://tec.mx/es/aviso-privacidad-participantes-expositores-panelistas-conferencias-moderadores" className="text-purple-400 hover:underline">
+              Aviso de Privacidad
+            </a>
+          </p>
+          <div className="mt-2">
+            <Checkbox
+              label=""
+              color="purple"
+              checked={formData.privacy_accepted}
+              onChange={(checked) => setFormData((prev) => ({ ...prev, privacy_accepted: checked }))}
+            />
+          </div>
+        </div>
 
-        // Staff Report Coordinator validation
-        if (formData.staffCoordinator.name) {
-            if (!formData.staffCoordinator.email)
-                newErrors.push('El correo electrónico de la coordinadora de informes (staff) es obligatorio');
-            else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(formData.staffCoordinator.email))
-                newErrors.push('El correo electrónico de la coordinadora de informes (staff) debe ser válido');
-            if (!formData.staffCoordinator.phone)
-                newErrors.push('El celular de la coordinadora de informes (staff) es obligatorio');
-        }
+        {/* Submit Button */}
+        <div className="mt-6 flex justify-end">
+          <Button
+            label="Enviar Registro"
+            variant="success"
+            showRightIcon
+            IconRight={withIconDecorator(Send)}
+            className="px-6 py-2 rounded-full flex items-center"
+          />
+        </div>
+      </div>
 
-        // Participants Report Coordinator validation
-        if (formData.participantsCoordinator.name) {
-            if (!formData.participantsCoordinator.email)
-                newErrors.push('El correo electrónico de la coordinadora de informes (participantes) es obligatorio');
-            else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(formData.participantsCoordinator.email))
-                newErrors.push('El correo electrónico de la coordinadora de informes (participantes) debe ser válido');
-            if (!formData.participantsCoordinator.phone)
-                newErrors.push('El celular de la coordinadora de informes (participantes) es obligatorio');
-        }
-
-        // Privacy notice
-        if (!privacyAccepted) newErrors.push('Debes aceptar el aviso de privacidad');
-
-        return newErrors;
-    };
-
-    // Handle form submission
-    const handleSubmit = async (e: React.FormEvent) => {
-        e.preventDefault();
-
-        // Run client-side validation
-        const validationErrors = validateForm();
-        if (validationErrors.length > 0) {
-            setErrors(validationErrors);
-            setSuccess(null);
-            return;
-        }
-
-        try {
-            // Convert files to base64
-            const profileImageBase64 = profileImage ? await fileToBase64(profileImage) : null;
-            const logoBase64 = logo ? await fileToBase64(logo) : null;
-            const participationFileBase64 = await fileToBase64(participationFile!);
-
-            // Prepare request body
-            const body = {
-                name: formData.venue.name,
-                location: formData.venue.location,
-                address: formData.venue.address,
-                logo: logoBase64,
-                participation_file: participationFileBase64,
-                generalCoordinator: {
-                    ...formData.generalCoordinator,
-                    profileImage: profileImageBase64,
-                },
-                associatedCoordinator: formData.associatedCoordinator,
-                staffCoordinator: formData.staffCoordinator,
-                participantsCoordinator: formData.participantsCoordinator,
-            };
-
-            // Get auth token (assuming it's stored in localStorage)
-            const token = localStorage.getItem('token');
-
-            // Make API request
-            const response = await fetch('http://localhost:3000/api/venues', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                    Authorization: `Bearer ${token}`,
-                },
-                body: JSON.stringify(body),
-            });
-
-            const data = await response.json();
-
-            if (!response.ok) {
-                if (response.status === 422 && data.errors) {
-                    // Handle validation errors from the backend
-                    const backendErrors = data.errors.map((err: any) => err.msg);
-                    throw new Error(backendErrors.join(', '));
-                }
-                throw new Error(data.message || 'Error al registrar el venue');
-            }
-
-            setSuccess(data.message || 'Venue registrado exitosamente');
-            setErrors([]);
-
-            // Reset form
-            setFormData({
-                generalCoordinator: {
-                    name: '',
-                    lastNameP: '',
-                    lastNameM: '',
-                    email: '',
-                    phone: '',
-                    gender: 'Mujer',
-                    username: '',
-                    password: '',
-                    confirmPassword: '',
-                },
-                associatedCoordinator: {
-                    name: '',
-                    lastNameP: '',
-                    lastNameM: '',
-                    email: '',
-                    phone: '',
-                },
-                staffCoordinator: {
-                    name: '',
-                    lastNameP: '',
-                    lastNameM: '',
-                    email: '',
-                    phone: '',
-                },
-                participantsCoordinator: {
-                    name: '',
-                    lastNameP: '',
-                    lastNameM: '',
-                    email: '',
-                    phone: '',
-                },
-                venue: {
-                    name: '',
-                    location: 'Puebla',
-                    address: '',
-                },
-            });
-            setProfileImage(null);
-            setLogo(null);
-            setParticipationFile(null);
-            setPrivacyAccepted(false);
-            setChosenSede(null); // Reset chosenSede on successful submission
-            setChosenGrupo(null); // Reset chosenSede on successful submission
-        } catch (err: any) {
-            setErrors([err.message]);
-            setSuccess(null);
-        }
-    };
-
-    return (
-        <form className='pagina-formulario' onSubmit={handleSubmit}>
-            <Navbar />
-            <div className="pagina-formulario min-h-screen bg-gray-900 text-white p-4 md:p-8 flex flex-col justify-center items-center">
-
-                <div className="info-formulario w-full max-w-6xl rounded-lg shadow-lg p-6 md:p-8">
-
-                    {/* Header */}
-                    <div className="flex justify-between items-center mb-6">
-                        <div className="flex items-center">
-                            <div className="w-2 rounded-full h-16 mr-4 notification-icon-purple"></div>
-                            <h1 className="text-2xl"><span className='italic'>Formulario de Registro</span><br /><span className='font-bold text-3xl'>Participantes</span></h1>
-                        </div>
-                        <Button
-                            label=""
-                            variant="error"
-                            showLeftIcon
-                            round
-                            IconLeft={X}
-                            onClick={() => console.log('Regresar clicked')}
-                            className="px-4 py-2 rounded-full flex items-center"
-                        />
-                    </div>
-
-                    {/* Section: Datos Personales */}
-                    <div className="mb-6">
-                        <h2 className="text-xl md:text-2xl font-semibold flex items-center mb-2">
-                            <span className=" mr-2"><DecoratedFlowerLotus width={25} height={25} strokeWidth={0} /></span> Datos del Participante
-                        </h2>
-                        <p className="text-gray-400 text-sm md:text-base mb-4">
-                            Responde con sinceridad las siguientes preguntas acerca de tus datos personales y de contacto. <br />
-                            Las secciones que contienen un asterisco (*) deben de responderse de manera obligatoria.
-                        </p>
-                        <p className="text-gray-400 text-sm italic">
-                            Es importante resaltar que este programa va dirigido a mujeres en secundaria o preparatoria / bachillerato interesadas en aprender y desarrollarse en áreas relacionadas con STEAM.
-                        </p>
-                    </div>
-
-                    {/* Form Fields: Datos Personales */}
-                    <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-                        {/* Nombre */}
-                        <InputField
-                            label="Nombre(s)*"
-                            placeholder="Edna"
-                            variant="primary"
-                            icon={"Fingerprint"}
-                            value={formData.generalCoordinator.name}
-                            onChangeText={(value: string) =>
-                                handleInputChange('generalCoordinator', 'name', value)
-                            }
-                        />
-
-                        {/* Apellido Paterno */}
-                        <InputField
-                            label="Apellido Paterno*"
-                            placeholder="Moda"
-                            variant="primary"
-                            icon={"Fingerprint"}
-                            value={formData.generalCoordinator.lastNameP}
-                            onChangeText={(value: string) =>
-                                handleInputChange('generalCoordinator', 'lastNameP', value)
-                            }
-                        />
-
-                        {/* Apellido Materno */}
-                        <InputField
-                            label="Apellido Materno"
-                            placeholder="Apellido Materno"
-                            variant="primary"
-                            icon={"Fingerprint"}
-                            value={formData.generalCoordinator.lastNameM}
-                            onChangeText={(value: string) =>
-                                handleInputChange('generalCoordinator', 'lastNameM', value)
-                            }
-                        />
-
-                        {/* Correo Electrónico */}
-                        <InputField
-                            label="Correo Electrónico*"
-                            placeholder="ednamoda@disney.com"
-                            variant="accent"
-                            icon={"At"}
-                            value={formData.generalCoordinator.email}
-                            onChangeText={(value: string) =>
-                                handleInputChange('generalCoordinator', 'email', value)
-                            }
-                        />
-
-                        {/* Celular */}
-                        <Dropdown
-                            label="Grado*"
-                            options={['1º', '2º', '3º']}
-                            value={formData.generalCoordinator.gender}
-                            onChange={(value: string) =>
-                                handleInputChange('generalCoordinator', 'gender', value)
-                            }
-                            variant="accent"
-                            Icon={withIconDecorator(GraduationCap)}
-                        />
-
-                        {/* Sexo */}
-                        <Dropdown
-                            label="Escolaridad*"
-                            options={['Secundaria', 'Preparatoria']}
-                            value={formData.generalCoordinator.gender}
-                            onChange={(value: string) =>
-                                handleInputChange('generalCoordinator', 'gender', value)
-                            }
-                            variant="accent"
-                            Icon={withIconDecorator(BookOpenText)}
-                        />
-                    </div>
-
-                    <div className="my-12">
-                        <h2 className="text-xl md:text-2xl font-semibold flex items-center mb-2">
-                            <span className=" mr-2"><DecoratedSealWarning width={25} height={25} strokeWidth={0} /></span> Datos del Tutor
-                        </h2>
-                        <p className="text-gray-400 text-sm md:text-base mb-4">
-                            Pídele a tu tutor que llene la siguiente parte del formulario con sus datos personales. <br />
-                            Las secciones que contienen un asterisco (*) deben de responderse de manera obligatoria.
-                        </p>
-                    </div>
-
-                    <div className="grid grid-cols-1 md:grid-cols-3 gap-4 mb-4">
-                        {/* Institución Académica */}
-                        <InputField
-                            label="Nombre(s)*"
-                            placeholder="Roberto"
-                            variant="primary"
-                            icon={"Fingerprint"}
-                            onChangeText={(value: string) =>
-                                handleInputChange('generalCoordinator', 'username', value)
-                            }
-                        />
-
-                        {/* Carrera */}
-                        <InputField
-                            label="Apellido Paterno*"
-                            placeholder="López"
-                            variant="primary"
-                            icon={"Fingerprint"}
-                            onChangeText={(value: string) =>
-                                handleInputChange('generalCoordinator', 'username', value)
-                            }
-                        />
-
-                        {/* Semestre */}
-                        <InputField
-                            label="Apellido Materno*"
-                            placeholder="Juárez"
-                            variant="primary"
-                            icon={"Fingerprint"}
-                            onChangeText={(value: string) =>
-                                handleInputChange('generalCoordinator', 'username', value)
-                            }
-                        />
-                    </div>
-
-                    <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                        {/* Locación */}
-                        <InputField
-                            label="Correo Electrónico*"
-                            placeholder="ejemplo@correo.com"
-                            variant="accent"
-                            icon={"At"}
-                            onChangeText={(value: string) =>
-                                handleInputChange('generalCoordinator', 'username', value)
-                            }
-                        />
-
-                        <InputField
-                            label="Celular*"
-                            placeholder="Puebla"
-                            variant="accent"
-                            icon={"Phone"}
-                            onChangeText={(value: string) =>
-                                handleInputChange('generalCoordinator', 'username', value)
-                            }
-                        />
-
-                    </div>
-
-                    {/* Section: Datos Coordinadora Asociada */}
-                    <div className="mt-8">
-                        <h2 className="text-xl md:text-2xl font-semibold flex items-center mb-2">
-                            <span className="mr-2"><DecoratedAddressBook width={25} height={25} strokeWidth={0} /></span> Selección de Grupo
-                        </h2>
-                        <p className="text-gray-400 text-sm md:text-base mb-4">
-                            Selecciona el grupo que más te llame la atención. <br />
-                            Puedes ver los detalles del grupo usando los botones del lado derecho.
-                        </p>
-                    </div>
-
-                    <div className='flex flex-col gap:4 lg:flex-row lg:gap-24 justify-center items-center mx-auto w-[80%] h-[65px] md:h-[50px] rounded-t-[15px] mt-8 input-secondary text-xs sm:text-base lg:text-xl'>
-                        <div className='flex'>
-                            <DecoratedStudent width={25} height={25} strokeWidth={0} />
-                            <h2 className='mx-4 font-semibold'>SEDE Elegida: </h2>
-                            <p>{chosenSede || 'Ninguna'}</p>
-                        </div>
-
-                        <div className='flex'>
-                            <DecoratedBookmarksSimple width={25} height={25} strokeWidth={0} />
-                            <h2 className='mx-4 font-semibold'>Grupo Elegido: </h2>
-                            <p>{chosenGrupo || 'Ninguno'}</p>
-                        </div>
-                    </div>
-                    <div className="fondo-tabla-forms flex flex-col p-6 gap-4 overflow-auto h-[50vh] sm:h-[75vh]">
-
-                        {/* Fila de búsqueda, filtro y botón */}
-                        <div className="flex flex-wrap justify-between gap-4">
-                            <div className="flex flex-1 gap-4 top-0">
-                                <div className="basis-2/3">
-                                    <InputField
-                                        label=""
-                                        showDescription={false}
-                                        placeholder="Search"
-                                        showError={false}
-                                        variant="primary"
-                                        icon="MagnifyingGlass"
-                                        value={inputValue}
-                                        onChangeText={(val) => setInputValue(val)}
-                                    />
-                                </div>
-
-                                <div className="basis-1/3">
-                                    <FiltroEvento
-                                        disableCheckboxes
-                                        label="Modalidad"
-                                        showSecciones
-                                        labelSecciones="Seleccionar"
-                                        secciones={modalidadesOptions}
-                                        seccionActiva={section}
-                                        onChangeSeccion={sectionFilterChange}
-                                    />
-                                </div>
-                            </div>
-                        </div>
-
-                        {/* Tabla */}
-                        <div className="overflow-auto">
-                            <table className="min-w-full text-left text-sm">
-                                <thead className=" sticky top-0 fondo-titulos-tabla text-purple-800 font-bold">
-                                    <tr className='texto-primary-shade'>
-                                        <th className="p-2 text-center">Grupo</th>
-                                        <th className="p-2 text-center">Modalidad</th>
-                                        <th className="p-2 text-center">Sede</th>
-                                        <th className="p-2 text-center">Cupo</th>
-                                        <th className="p-2 text-center">Horarios</th>
-                                        <th className="p-2 text-center"></th>
-                                    </tr>
-                                </thead>
-                                <tbody className="text-gray-700">
-                                    {paginatedData.map((grupo, index) => (
-                                        <tr key={index} className="border-t border-gray-300">
-                                            <td className="p-2 text-center">{grupo.grupo}</td>
-                                            <td className="p-2 text-center">{grupo.modalidad}</td>
-                                            <td className="p-2 text-center">{grupo.sede}</td>
-                                            <td className="p-2 text-center">{grupo.cupo}</td>
-                                            <td className="p-2 text-center">
-                                                {grupo.horarios} <br />
-                                                {grupo.fechas}
-                                            </td>
-                                            <td className="p-2 flex gap-2 justify-center">
-                                                <Button label='' variant="success" round showLeftIcon IconLeft={Check} onClick={() => handleSelectClick(grupo)} />
-                                                <Button label='' variant="primary" round showLeftIcon IconLeft={Eye} />
-                                                <Button label='' variant="warning" round showLeftIcon IconLeft={Star} />
-                                            </td>
-                                        </tr>
-                                    ))}
-                                </tbody>
-                            </table>
-                        </div>
-
-                        {/* Paginación */}
-                    </div>
-
-                    {/* Section: Datos Coordinadora de Informes (Staff) */}
-                    <div className="mt-8">
-                        <h2 className="text-xl md:text-2xl font-semibold flex items-center mb-2">
-                            <span className="mr-2"><DecoratedHeart width={25} height={25} strokeWidth={0} /></span> Permiso de participación
-                        </h2>
-                        <p className="text-gray-400 text-sm md:text-base mb-4">
-                            Dentro de esta sección tendrás que subir tu permiso de participación, la cual deberá de estar firmado por tu tutor. <br />
-                            Esta sección es obligatoria.
-                        </p>
-                    </div>
-
-                    <div className="mt-6 p-4 bg-white text-black rounded-lg tarjeta-archivo">
-                        <div className="flex items-center titulo-tarjeta-archivo">
-                            <span className=" text-2xl mr-2 icono-tarjeta-archivo"><DecoratedFilePdf width={25} height={25} strokeWidth={0} /></span>
-                            <h3 className="text-lg font-semibold">Sube tu permiso firmado</h3>
-                        </div>
-                        <p className="text-sm my-6">
-                            Selecciona un documento para subir.  Ten cuidado al subir tus documentos y verifica dos veces que se suba correctamente.
-                        </p>
-                        <input
-                            type="file"
-                            accept=".pdf"
-                            onChange={(e) => handleFileChange(e, setParticipationFile)}
-                            className="mt-4"
-                        />
-                        {participationFile && (
-                            <p className="mt-2 text-xs text-gray-600">
-                                Archivo seleccionado: {participationFile.name}
-                            </p>
-                        )}
-                    </div>
-
-                    {/* Aviso de Privacidad */}
-                    <div className="mt-8">
-                        <h2 className="text-xl md:text-2xl font-semibold flex items-center mb-2">
-                            <span className="mr-2"><DecoratedMegaphone width={25} height={25} strokeWidth={0} /></span> Aviso de Privacidad
-                        </h2>
-                        <p className="text-gray-400 text-sm">
-                            Confirma que he leído, entendido y acepto el Aviso de Privacidad disponible en:<br />
-                            <a href="https://tec.mx/es/aviso-privacidad-participantes-expositores-panelistas-conferencias-moderadores" className="text-purple-400 hover:underline">
-                                https://tec.mx/es/aviso-privacidad-participantes-expositores-panelistas-conferencias-moderadores
-                            </a>
-                        </p>
-                        <div className="mt-2">
-                            <Checkbox
-                                label=""
-                                color="purple"
-                                checked={privacyAccepted}
-                                onChange={setPrivacyAccepted}
-                            />
-                        </div>
-                    </div>
-
-                    {/* Submit Button */}
-                    <div className="mt-6 flex justify-end">
-                        <Button
-                            label="Enviar Registro"
-                            variant="success"
-                            showRightIcon
-                            IconRight={withIconDecorator(Send)}
-                            className="px-6 py-2 rounded-full flex items-center"
-                        />
-                    </div>
-                </div>
-            </div>
-        </form>
-    );
+      <Modal
+        isOpen={isErrorModalOpen}
+        onClose={() => setIsErrorModalOpen(false)}
+        title="Errores en el formulario"
+        messages={errors}
+      />
+      <Toast
+        isOpen={isSuccessToastOpen}
+        onClose={() => setIsSuccessToastOpen(false)}
+        message={success || ''}
+      />
+    </form>
+  );
 };
 
 export default ParticipantRegistrationForm;
